Change os-ansible-deployment checks to use scripts
This commit changes the os-ansible-deployment checks to exclusively
make use of scripts which are located in the repository which is
cloned. This is done to improve flexibility - if the check needs to
change, it can be done within the script instead of submitting a review
to the project-config.

This commit also implements the commit check as a voting check across
all branches, and also changes the merge check from the lint check to
the commit check.

Change-Id: I95d4f12e374b80be70f513429a5dc04b0bdf3534