Remove wasteful tripleo workspace-cache cloning
There doesn't appear to be anything in TripleO that relies
on these clones. Also, this is a bit confusing since people
may expect that things copied into ~workspace-cache would
eventually end up at /opt/git/ (this is not the case).

The correct way to add projects to a job is to export a
PROJECTS variable like devstack-gate does.

Change-Id: I0a96855ed9a38bcade6e506b37621b6b7b348b49