Ironic: Bump memory size for non-concurrency job
In Ironic we want to be able to use the IPA ramdisk as the default ramdisk
for our deployments. Since that ramdisk is a heavier than our previous one
we need to bump the memory size from 512MB to 1GB to be able to test it
in gate. It's important to note that since the concurrency job needs 3 VMs
we will still run it using the old ramdisk until we can solve the problem.

Plus, for normal jobs where we bumped the RAM size we are lowering the
number of VMs created from 3 to 1, because only 1 VM get's used anyway.

This patch is making use of the new DEVSTACK_LOCAL_CONFIG that allows
passing arbitrary config options to devstack-gate.

Change-Id: Icf011d995640cabcb3677f814b7dd37aab0a303b