Fix swift func tests to run using both auth versions
The swift functional test job was changed to run
tox -e func as sudo [1]. The job first runs functional tests
against keystone, then repeats the tests against tempauth.
The second part of the job requires SWIFT_TEST_CONFIG_FILE
env var to be set to cause the tests to run against tempauth.
The introduction of sudo in [1] causes this env var to be lost,
so the tests run against keystoneauth again.

A tell-tale is that the second run of the functional tests should
output "SKIPPING FUNCTIONAL TESTS SPECIFIC TO AUTH VERSION 3" before
the first line of test results. This has been missing since [1] was
merged.

[1] change id Ifc3e862b0fe86928c3943a41f07cd086d6ace6fe

Change-Id: Ibea153e94e7919b5422a344d365cee3106efb2fe