Rename functional files to projects.
Generally this area of the repo is organized by project, so rename
the *-functional.yaml files to avoid the suggestion that each job
needs its own file.  Instead, name them for their projects to suggest
that you can look in each project file to see the jobs that apply
specifically to that project.

Change-Id: I31e95a405c03c207e26f9a63403e1b7205595b7f