Use neutron's api job to validate func job change
A recent commit to devstack-gate added support for limiting the repos
that are updated for a node by overriding the PROJECTS variable.  This
change temporarily repurposes neutron's api job to check that setting
this override is compatible with the functional job.  The api job is
currently non-voting, so no patches will be harmed in the process.

Change-Id: I7591569d3f039eeebed859333b29dd63104cd389