Revert "Remove lib-forward-testing from openstackclient"
I believe this was erroneously removed, the changes that should have
been made have been made in https://review.openstack.org/#/c/158045/

We wanted to get rid of the juno/icehouse jobs, not the one that
tests off of master.

As of now, osc is just performing python-related jobs and a functional
test. We need a DSVM test.

This reverts commit 33a6babe6a5a1dcdabb701bd986a239ecb9daef3.

Change-Id: Id01f0c9bcf246ab7263e78231bbb442d38652b69