Create our own rally job for ec2-api
In previous review
https://review.openstack.org/#/c/159138/
I tried to use existing rally job. But I skipped fact that in this job
ec2-api wouldn't loaded. Now I was trying to think up how this existing
job can be adapted to load ec2-api plugin. And simple way to do it -
is to create another job-template with needed shell script and needed params.

Change-Id: I3a95b298a70795601ef4002fc413e91960af0f2b