Fix glob pattern for os-ansible-logs publisher
Neither Ant's nor infra's documentation is clear, but it appears that
keep-hierarchy shouldn't be true when also specifying a pattern that
only looks for filenames that match a pattern. Instead, just copy all of
the sub-directories and their contents within /openstack/log to ensure
all log files are copied during os-ansible-deployment gate-commit-check
jobs.

Change-Id: I0b3d8c0b81a1cd26806be824843290f4153daee0