Fix swiftclient func tests to run using both auth versions
The rationale for this change is the same as for merged change
to swift func tests job [1].

The swiftclient functional test job was changed to run
tox -e func as sudo [2]. The job first runs functional tests
against keystone, then repeats the tests against tempauth.
The second part of the job requires SWIFT_TEST_CONFIG_FILE
env var to be set to cause the tests to run against tempauth.
The introduction of sudo in [2] causes this env var to be lost,
so the tests run against keystoneauth again.

[1] change id Ibea153e94e7919b5422a344d365cee3106efb2fe
[2] change id Ifc3e862b0fe86928c3943a41f07cd086d6ace6fe

Change-Id: Ieea3efbd6526edb8038b677a2e2278107acfe211