Disable gate-oslo-incubator-python34 for Icehouse
The job was broken by recent switch from py33 to py34 jobs. Precise
images do not contain latest python version, and fail now.

As per oslo team,

"The incubator python 3 job was added to help us port incubated code to
python 3 before graduating it. We won't be graduating modules from the
stable branch, so as long as none of the consuming projects have python
3 jobs on their stable branches we can just drop the icehouse python 3
job for oslo-incubator."

I don't know of any consumers that actually use code from oslo-incubator
and at the same time support py3, so disabling the job for the branch.

Change-Id: Ia8f5fac1d37a0ce8062ba144da9c3adcf9a474a3