Drop ironic tempest regex, stop running all of Tempest
This drops the ugly tempest regex we were using to target the subset of tests
we knew to be functional against baremetal, and instead uses the 'baremetal'
service tag to run only relevant tempest tests.

This updates the parallel and grenade jobs to do the same. Previously, these
jobs were running the entire Tempest suite against baremetal. This takes a
very long time and runs many tests that are not really relevant.

This is a shift in a previous effort to enable running the entire Tempest suite
against Ironic to running only the smallest subset of tests we care about.

Also, this enables tenant isolation where it is disabled, since that is now
functional against Ironic.

Change-Id: I1a979c644b4950f0fce59c764b115266b8a13ca8