Add gate-tempest-dsvm-full-trove
Because trove builds on the compute base layer, it does not need
to be tested in all devstack-gate tests.  Instead, create a new
job that tests trove in a devstack environment as a first step
in removing it from the standard configuration.

Specify that TROVE should be enabled in the devstack updown job
so that the parts that enable trove are exercised in devstack.
This is already currently the case, but this updates the job config
so that it will continue to do so later when we remove trove from
the default config.

Change-Id: Ib503e45b08edfb2252d7c56af14f97d630e137d8