Partial revert of func job setup for VPNaaS repo
As part of enabling a new experimental functional test job for the
neutron-vpnaas repo, file filtering was incorrectly added to the
existing OpenSwan and new StrongSwan functional jobs.

The original intent was to have the StrongSwan job run tests in the
parent directory (shared tests) and the strongswan sub-directory, and
to have the OpenSwan job run tests in the parent directory and the
openswan sub-directory.

This wrong on two counts. First, these file filters are used to
determine what files changed will *trigger* the job. As it currently
stands, these filters are detecting only functional test file changes
and not neutron-vpnaas production code changes.

Second, to get the jobs to run on the desired set of functional test
files, shared test modules will be put in a sub-directory, and not in
the parent directory. Changes to tox.ini and hook files in the repo
are being used to select the right directories for testing.

Change-Id: I214f2c9610aedac2192d6bb777cd485688f938c6
Partial-Bug: 1424757