Ironic: Move pxe_ipa-nv to check instead of experimental
The pxe_ipa-nv job is testing the future default ramdisk for Ironic, at
first we added it to experimetal just to make sure it works. Now that
we know it's passing we are moving it to the "check" so it can run on
every gate check for us to collect data and make it voting in the future.

Change-Id: I7187bdb1480cafa2f8152d0b0229a436c1fa3514