VPNaaS make new functional test non-voting
The experimental job for the new StrongSwan based functional test
for the neutron-vpnaas repo is being moved to the check queue as a
non-voting job, so that it runs on all commits.

This job sets up DevStack to use the new StrongSwan driver (to be
reference implementation in future).

Change-Id: I33fbe92385b5ed8da6b525127f0c669f16c2dcee
Partial-Bug: 1424757