Set experimental pipeline precedence to 'normal'
Running the experimental pipeline at a 'low' precedence has resulted
in job maintainers adding new jobs to the check pipeline as non-voting
to ensure reasonable turnaround.  This costs a node per patch under
test, though, and it's preferable to improve the turnaround time of
the experimental pipeline by matching its precedence to that of the check pipeline.

Change-Id: I058e4035b04c15c3fa4f6c48eecb606dda765187