project-config: Merge lint jobs (2/2)
Since both pep8 and bashate are voting, we can merge the two lint
targets into the "common" pep8 lint target. There's no need for a
separate bashate run.

This change is the second part for this.
Since the jobs are changed, we can now remove the obsolete bashate
tox.ini target.

Change-Id: I572ba3fe8b3a9a05f5239a858d2f701bfd8f684d