Create jobs for a wheel mirror
In order to speed up tests where we install python packages we want to
create wheels for packages we depend on.

This patch creates four new scripts:

* afs-slug, which generates a string version of the platform architecture,
  derived from the host system's parameters.
* wheel-build, which generates all wheels, from all branches, in `pwd`/wheels
* wheel-copy, which copies this wheels directory into our AFS file structure.
* wheel-index, which rebuilds the index.html file from the directories
  found in the afs directory.

The above scripts are executed in sequence, with afs credentials only loaded
where necessary. Additional timeouts have been provided to prevent things from
locking.

This patch assumes that a mod_rewrite rule will be available on the serving nodes
that unmunges the URL.

Co-Authored-By: Sean Dague <sean@dague.net>
Depends-On: Ie33861dbcb413c34012820fda76f8be94fb1d151
Depends-On: I81c39d420d8ac70def57949ea0d4c323b8797086
Change-Id: I3463fba889239c565c86ffb080adffd7bfdd18f9