Don't dump Jenkins proxy errors to console logs
When grabbing the console logs for jobs in order to upload these logs to
swift we may run into Jenkins proxy errors which then end up in our
console log file.

Avoid this by using curl --fail which will not output any data and
return 22 if the request fails.

While editing this curl command document the odd choice of flags given
as their necessity may not be very apparent.

Change-Id: I7e53e11882046282b95addd635d804fa29b0eb44