Disable -dibtest jobs
The -dibtest jobs are experimental jobs against devstack-gate to test
dib images for trusty and precise as we bring those images up in new
regions. We are not currently able to bring those images up in rackspace
and we need to free up some disk space on the nodepool server so we are
cleaning up those images. To prevent zuul from getting cranky we also
remove the jobs that use these images from zuul.

Change-Id: I91f8d91318752e09a15d64c8b3c506cf83badca2