Adjust regression exceptions for Nova Cells V1 job
Since several patches have merged, the set of tempest test expected
failures with cells have changed. This patch adjusts the list of
tests that should not be run in order for the job to pass.

The tests being removed from the blacklist had been failing in the
past because a flavor created by tempest could not be accessed in
child cells during instance create. Patches have since merged that
enable use of the flavor object stored in the instance object,
eliminating the need for a flavor lookup from the database. So tests
having to do only with an instance create can be removed from the list.

There are some classes of tempest tests still expected to fail in a
Cells V1 environment:

* Networks API
* Resize instance
* Floating ips
* Shelve instance
* Rescue instance

Change-Id: If32b7361521d4747dcd05db47e3bf4955ce1b232