Improve ec2-api gating jobs for rally testing
1. Enable tempest for rally gate. Only tempest preparation code has needed falvors creation -
flavors without disk(m1.nano for example). We need flavor without disk because only this can be
run with fakevirt driver due to bug https://bugs.launchpad.net/nova/+bug/1433994

2. Add calling hook script before running rally to setup specific things of third-party project.
For example - botocore. This library is absent in global requirements and can not be add to ec2-api
project. And global requirements is frozen now for adding new.

Change-Id: I9993c97ecb6900efefc8ae73107d8a06a9e6dcf6