Add os-testr repo
This commit adds a new repo, os-testr. The intent of this repo is to
provide tooling to interact and use testr. Right now it contains 2
pieces subunit-trace, which has become the de-facto output filter for
running tests, and ostestr which is a testr wrapper similar to the
pretty_tox.sh scripts which are currently used everywhere but expands
on that functionality.

Change-Id: I994b1271b69741161d1ce165eda4ae32d529bdc4
Depends-On: Idbd6251546809f6ea7a651be9e758619fddd4e7f