Fix destination for requirements clone dest
In the requirements checking job we need to clone the requirements repo
into the tmpdir that the test creates for the repo. Otherwise subsequent
dir changes end up failing because the dir they attempt to change into
fails.

While we are at it print the stdout and stderr written by zuul cloner to
aid in debugging.

Change-Id: Id39140ac2a0ccbaf9a9ac0f6f76350c539048670