Enable trove plugin for trove tempest job
Trove will no longer be tested in the devstack updown job because
its configuration has moved from the devstack repo to the trove
repo, so remove the no-longer-necessary enable_trove flag in that
job.

Also remove the flag from the tempest job but replace it with the
configuration option to enable the plugin so that job does not
functionally change.

Change-Id: Iac738fb6071c8cb60961872cdef371582041d0a2
Depends-On: I8f9f1a015edb7ec1033e2eaf0b29ab15d89384ce