Gnocchi: do not check requirements
As the Gnocchi project does not follow the integrated release schedule,
we are going to release on a different pace than most OpenStack
integrated projects. Having the requirements tight to the rest of the
schedule might very likely be an issue for deployer wanting to upgrade
Gnocchi while the rest of OpenStack isn't released yet.

So Gnocchi is going to follow its own dependency rules, while doing best
effort to be compatible with the rest of OpenStack obviously.

Change-Id: Id2836cf52b62ae22d39135ee725679c68fda06cf