Optimise MidoNet devstack-gate job
Since MidoNet only implements the Neutron API, we can leverage the
OVERRIDE_ENABLED_SERVICES env var in the devstack-gate to only enable
the minimal set of services to provide Neutron features and optimise the
time and resources the job needs to run.

Set also the minimal set of tests to guarantee that new commits don't
break the gate through the DEVSTACK_GATE_TEMPEST_REGEX env var.

Change-Id: I7a5838fd7d316c0f3b88b388e797171b4a6e3ff3