Stop running gate-tempest-dsvm-neutron-src-python-glanceclient-juno
These 2 jobs were running on python-glanceclient for all branches:
- gate-tempest-dsvm-neutron-src-python-glanceclient
- gate-tempest-dsvm-neutron-src-python-glanceclient-juno

Currently the -juno version is blocking glanceclient development due
to a version mismatch (bug 1442682).

Stop running the -juno version: the client is currently capped in global
requirements for stable/juno, and there will be a stable/juno branch for
glanceclient soon, so updates to the glanceclient master branch should
not affect stable/juno.

This is the glanceclient equivalent of this (already merged)
keystoneclient change: https://review.openstack.org/#/c/172662

Change-Id: Ia1b552b935a37a0a19654a3e1d90e22bc485c187
Closes-bug: 1442682