change name from aiopcpu to multinode
aiopcpu is an exceptionally cryptic name that doesn't mean anything to
most people reading test results. multinode is what the job is, and is
much more understandable at first glance.

This changes the naming from aiopcpu => multinode. Should we create a
different kind of multinode test in the future, we can come up with an
appropriate name for it then.

This is safe to change because the devstack-gate logic actually is
fully != 'aio', not == 'aiopcpu'.

Change-Id: Ic7fe31a397af8a14ecf33a4902908ff9c03d11d3