Temporarily copy proposed/* ACLs to stable/kilo
The proposed/kilo branches of integrated release projects will not
be used this cycle. Instead we're going straight into stable/kilo
but keeping the access controls we had over the old proposed
branches for now. This change can be reverted once Kilo is released,
to return control to teh stable branch maintainers thereafter.

Change-Id: Iaced68397d700d19a3f715de458bd83cbf78cf39