Add puppet stackforge modules to tripleo jobs
We need to set the PROJECTS variable so that it contains all the
various puppet stackforge projects so that we can have proper
refspecs setup in /opt/stack/new for our CI jobs. This will
allow TripleO to provide CI coverage on these projects
once we decide to enable it.

Change-Id: I0aced3c0d1f3d2ce0bca00de847e41eee4dd2877