Handle empty translation files
We have some strange files in the tree that contain no or little
translation after merging - and these are downloaded from transifex.
So, these files are not known to git and git rm fails.
To not special case every git rm and git reset for both old and new
files, re-add a "git add" at the beginning so that we can operate on
files known to git.

Note: This happens when we have a file like glance-log-critical.po that has no
strings - but we had to create it with initial content in transifex,
since empty files are not allowed there.

Change-Id: Ie9bb3e59560d50829e56928b2ed79e3999dd39cb