OpenStack Foundation Transparency Policy project
The Transparency Committee has been using google docs to work on some
proposed changes to the OpenStack Foundation Transparency Policy. By
moving this process to git, we hope make the process more accessible
and to keep a number of logically distinct proposed changes separate
for individual review by the board.

The gerrit ACLs are modeled on the ACLs for the openstack/governance
repo. That is, anyone can leave a "Code Review" +1/-1 vote. Only
transparency-committee members can add a "Rollcall-Vote" +1/-1. And,
finally, only the transparency-committee-chair group can use +1/-1
Workflow to merge or mark as Work In Progress.

In the end, we may up with the foundation board chair (currently
Alan Clark) as the sole member of transparency-committee-chair, since
policy changes should not be merged until they have been approved by
the board.

Depends-On: Idcfa3162514fb4d02a49f20efbe43a8163f9b162
Change-Id: Ic0a4153623f3640da142253938c79bc4af26a8c5