beaker: support ssh service name on centos
In a previous patch, we added Centos support for Beaker but we forgot
to change the SSH service name.
On ubuntu, service name is ssh, and on RH systems it's sshd.

Also use systemctl on RH because 'service' is not in PATH:
/usr/bin/systemctl
/usr/sbin/service
Since we use sudo without option, we should use systemctl to manage the
service.

Also install `bundle` which is needed by Beaker to run.
This package was installed in the previous image we were using, but not
on this one.

Change-Id: I6be34f2c6a8b4effd49e27dff0fbdeffafc4c5cb