Cleanup openstack documentation Gerrit ACLs
api-site and openstack-manuals share the same permissions, use only one
config file.

Remove not needed "refs/heads/stable/" sections, these are not needed
with current gerrit project setup, we inherit from "refs/heads/" and the
permissions are the same.

This is a cleanup, there should not be any functional change for any of
these repositories.

Change-Id: Iba09f6ea8618740790ae51ecbf08010adb4f432e