make pypy non-voting across the board
This is breaking at a fundamental level on all jobs. We should disable
voting for pypy across the board to get things moving again, and
consider removing this in the future.

Consolidate the multiple definitions for pypy that would be created
with this, and place the comments next to the correct logical blocks
to make them less confusing.

Related-Bug: #1290562

Change-Id: If1cf7af0e950f9b7725e84127ee936bffd63ea6d