Set the Kolla build timeout to 90 minutes
We believe the Kolla build will take between 1 and 3 hours.  We
are not certain exactly how long it will take, so we plan to
start with 90 minutes and later tune the timer to match our
average run-times + 20%.

Note our gate functional job builds about 40 docker images which
takes quite awhile.

Change-Id: Ifd052464cacbe4ce4d0584277aa729ab151d48fe