Add keystoneauth-core group
To enable a wider team (including those from the SDK effort) to have
a say in how keystoneauth is managed as a library, add a dedicated
core group for both the library and release. This will allow for
keystoneauth to include cross-team core members without needing
to add these members to the core team that manages keystone server
code.

The plan is to also include the keystone-core and keystone-release
groups respectively.

Change-Id: I40721df75e86e7dcd3b99ad5a4e78ddcfa432645