Use zuul-cloner to clone boostrap dg repo
Before devstack-gate runs we have to clone devstack-gate then start it.
Previously this was a semi complicated JJB macro based on whether or not
the repo exists previously. It also was unaware of the git cache on our
nodes. We can simplify all this and use the git cache by having
zuul-cloner do this initial set up for us.

Change-Id: I512203f35797837e1f4612add05f60b759a6eb07