Reconfigure irc-meetings output
Update the irc meetings publish job to expect all output in the
output/ directory and copy the entire directory contents to
eavesdrop.o.o.

The previous configuration actually caused jenkins to create
subdirectories with the names of the files we intended to copy.

Also, rename the job to match existing patterns.

Change-Id: Ic188c1849559a9c8f282845940f56cf533b9bfd6