Install zlib1g-dev for puppet beaker jobs
The latest version of beaker-rspec depends on Nokogiri 1.6.x. We are
currently pinning beaker-rspec to 2.2.x which brings in Nokogiri 1.5.x.
Nokogiri 1.6.x has libxml2-dev and libxslt-dev vendored within the gem,
so once we unpin beaker-rspec we will no longer need to install those
packages.  However, on Debian/Ubuntu we still need to install the
header libraries which come in the zlib1g-dev package and which do not
seem to be included in libxml2-dev or libxslt-dev. If we do not, then
attempting to use the latest version of beaker-rspec results in gem
installation errors[1]. This patch adds the new package, zlib1g-dev, so
that both Nokogiri 1.5 and 1.6 will work, and therefore beaker-rspec
2.2 and latest will work. See the troubleshooting documentation[2] for
more information.

[1] http://logs.openstack.org/61/185161/1/check/gate-puppet-nova-puppet-beaker-rspec-dsvm-trusty/2152ed2/console.html.gz#_2015-05-22_21_53_47_219
[2] http://www.nokogiri.org/tutorials/installing_nokogiri.html#ubuntu___debian

Change-Id: Iefdf96325e11b470e864d5386a3eab0eb655c960