Ignore requirements job on keystoneauth branch
keystoneauth is extracting the reusable parts of keystoneclient into a
separate project that can be consumed by other clients. To handle the
migration we have a dependency on the git head of the keystoneauth repo
from the feature branch of keystoneauth which fails the requirements
check.

Turn off the requirements job until we are ready to release
keystoneauth.

Change-Id: I1af32951d511dd4fff9a8ebca05f6aea8ab2835c