Remove stable cross-test jobs from cliff
Remove the jobs that run the unit tests of cliff against stable OSC and
neutronclient. We rely on the dsvm jobs to run integration tests of
cliff changes, using both OSC and neutronclient for stable branches.

Configure the jobs that run the unit tests of cliff against stable OSC
and neutronclient to only run when patches are submitted to master.

Change-Id: I42c50ea8fd4b185dd1e745d3bdb6695814840e24