Expand the vote range for infra-specs
The voting configuration for infra-specs produces a red-x and
green-check for +/-1 code reviews now, which is confusing to people
since they normally associate those icons with +/-2 votes.  This
change expands the vote range to include +/-2 but gives no one
permission to vote at those levels.  This should cause Code-Review
votes to appear as +/-1, while the NoOp function continues to
mean that +2 votes are not necessary.

Also extend the Council-Vote to +2, but similarly give no permissions
to use it so that council votes will appear as +1, -1, or red-x in
the case of a -2 veto (the -1/green-check dichotomy would also
seem weird for similar reasons).

Change-Id: I042e6168334fb11291edbaed97b3ae2424218166