Use "Rollcall-Vote" for TC and infra-council
Adopt the idea that a review label name should be as
widely-applicable as possible for a given concept so that different
repos with similar needs will end up sharing columns in the user
interface.  Since the TC and infra-specs repos both need "votes from
the people who are members of the committee" as a review label
(while preserving the ability for non-members to leave their feedback
as +/- entries in code-review), have them share the name "Rollcall
Vote".  This will cause them both to appear in the "RV" column in
Gerrit change lists.  Simply "Vote" is an option worth considering,
but it is not clear how "Vote" is different from Code-Review.

Change-Id: Id53d1bea7f53b92009ba6e9274162b8ffa3b2b92