Add pre/post test hooks for Heat grenade job
To test out of tree grenade plugins, we'd need to set
environment variables before hand to instruct grenade
on where to find the plugins and load it's settings.

Note:- This is going to make the current non-voting
grenade job fail because the directory structure doesn't
exist in Heat but it should happily pass should the
out-of-tree plugin work.

Change-Id: I7b548cb8c7f07735d1b88d7cb4e4d96e961cc591