networking-ovn: Allow setting devstack-gate env vars
I recently added the ability to have pre and post gate hook scripts in
networking-ovn.  What I really wanted this for was to be able to set
DEVSTACK_GATE_TEMPEST_REGEX in the pre hook script.  It turns out this
doesn't actually work like I expected.  It seems fine to leave the
hooks for future use, but add a new call to a script to setup some env
vars before calling devstack-gate.

Change-Id: I67dceb8087d60e1bfc9d6402edcbbe5f5026fcba
Signed-off-by: Russell Bryant <rbryant@redhat.com>