Make gate-cinder-pylint non-voting
The cinder-pylint check doesn't work with the proposal-bot
tag merges and causes said changes to be blocked.

This patch makes the pylint checks in Cinder non-voting
temporarily until we either fix things up and make it work
properly, or decide to remove it altogether.

Have opened LP# 1463530 against Cinder to address this.

Change-Id: Ie29d6dd42706b1833309c6af70951076685bbb96