puppet: make Puppet 4 unit test jobs voting
All Puppet OpenStack modules are Puppet4 compliant.
The jobs were not voting but since now all of our modules are ready, we
want to have the job voting so we keep the job stable and our code
Puppet 4 compliant.

It also makes sure it's not blocking stackforge/puppet-openstack-cloud
Puppet module, which is not part of Puppet OpenStack modules.

Change-Id: I961a64d5495b2126a25404e953c4681d40506c28