puppet-openstack: make vote beaker jobs
All our modules now have beaker tests.
That means jobs are green and we need them voting to keep our modules
stable and testable.

This patch enable voting on Beaker jobs for Puppet OpenStack modules
after Juno release (and still don't run them for Juno and before).
It also adds the jobs in the gate (it was only in check pipeline).

Change-Id: Icb09ef78123b1e8f9a2641af01b3aae04c44cf9b