Change tag ACLs for library projects to library-release team
As part of the plan to move release tagging permissions to a core group
managed under the release management team [1], update the ACLs files to
give permissions so a new library-release team instead of the existing
core or release teams. When the new team is created, the "Release
Managers" group and a small number of volunteers will be added to it.

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-June/066346.html

Change-Id: I9d1568fa688c08b658444b33df0408907d0b3fa6