Move puppet-murano from stackforge to openstack
This change moves puppet-murano from stackforge to openstack in
order to be included as part of OpenStack Puppet Modules project.
This repository will contain the module necessary to configure
and manage Murano with Puppet.

This change is related to a proposed governance change:
Id846941c9002194e76a6ac31c1979d410aa25fb9

Change-Id: Ia7e917b3032ef5ea56b33cf92e6b7c2d62674f6a