Remove mount of ephemeral disk
hpcloud has started sending metadata to cloud-init to mount ephemeral
disks.  This ends up writing a fstab entry for /dev/vdb.

---
$ curl http://169.254.169.254/2009-04-04/meta-data/block-device-mapping/ephemeral0
/dev/vdb
---

It's unclear why this just started happening, but it did.

devstack-gate later attempts to repartition this into swap & disk
space and mount it elsewhere.  So remove this mount from fstab -- it
shouldn't come back because the next thing we do here is tell
cloud-init to not use the metadata source.

Change-Id: I3787d0f7e5139e891686ffbb2970e65d09f112b1